+++ CHANGES.rst
@@ -22,6 +22,8 @@ number of rows, each row of `x` is filtered with the respective row of `b`.
 - The `Experiment` class tries to create the output folder if it does not exist.
 - The speech material name is saved out the output data frame when running a
 speech intelligibility experiment.
+- Added the function py:func:`~pambox.speech.material.Material.average_level`
+to measure the average level of a speech material.
 
 Performance
 -----------